CB-8595 Switch to single (un)installer function for platform handlers
Add getFilesAndFrameworks() to plugin info. The order of items returned by it
is the same as was in both install and uninstall.

Why: When (un)installers are accessed as

    handler = platforms[p]
    handler[fileType].install

it's more difficult to create a base class for the handler, the derived class
can't override only the [fileType].install but not the uninstall (search order
in the prototype etc). Moreover, most (un)installers are not to different for
different file types, iOS already uses a common installers that accepts the
file type as the first parameter.